UPLOAD Fortnite External Drivers
This is for all winver, it's the "may2h2drve" driver for Fortnite. Yh it's good i guess. And no it's not undetected. BRUH Don't fall for people saying that this is UD. LOL. This is not for pasting, and no. It's not mine. Since it's not posted anywhere. I decided to post it.

Credits: ritz (got the drivers from him)

Join our discord: https://discord.gg/YpHaVa5Bg5